requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
Flask-Login==0.6.3
fonttools==4.45.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
protobuf==4.21.12
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
Werkzeug==3.0.1
